requirements.txt
Flask==0.12.4
Flask-WTF==0.12
Flask-Script==2.0.5
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.1
alembic==0.9.6
psycopg2==2.6.1
WTForms==2.1
twilio==6.19.2
celery==4.2.0
SQLAlchemy==1.1.15
python_bcrypt==0.3.2
profanity==1.1
pygsheets==1.1.3
GroupyAPI==0.10.2
Flask-Scss==0.5
oauth2client
redis
